[FE_refactor] 자동 로그아웃 기능GlobalLayout으로 이동 및 alert 기능 제거
